# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ETGA API Listing",
        "description": "This is the List of APIs that should be used across ETGA and Tourguides ( Dashboard ETGA, Dashboard Tourguides, Website ETGA, Website Tourguides.me)",
        "contact": {
            "email": "nada@technologies.ae"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "List all documents",
                "operationId": "479e53af7d4820aecb8f9534fdbe21aa",
                "responses": {
                    "200": {
                        "description": "A list of documents",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Document"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Create a new document",
                "operationId": "9b247d60e8834d470666fba8e773edb4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Document"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Document created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/documents/{id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get a single document by ID",
                "operationId": "0fa5befcdb9042b346d01baea49dee95",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of a single document",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Documents"
                ],
                "summary": "Update an existing document",
                "operationId": "b851458865e322163fe9b88595919255",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Document"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Document"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Delete a document",
                "operationId": "889c107c27dc6b118dee932d1760f574",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Document deleted successfully"
                    },
                    "404": {
                        "description": "Document not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/documents/upload": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload a document",
                "operationId": "8e3a5d070f386f819b02061c0ef0adc1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "The file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "File uploaded successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/documents/list": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "List all uploaded files",
                "operationId": "1d32277c52f31b389ab70bc4b82e69bd",
                "responses": {
                    "200": {
                        "description": "A list of files",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "documents/file1.pdf"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Document": {
                "required": [
                    "type",
                    "file_path",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier for the document",
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "description": "The type or category of the document",
                        "type": "string",
                        "example": "Invoice"
                    },
                    "file_path": {
                        "description": "The path where the document file is stored",
                        "type": "string",
                        "example": "/uploads/documents/invoice_123.pdf"
                    },
                    "user_id": {
                        "description": "The ID of the user who owns or uploaded the document",
                        "type": "integer",
                        "example": 123
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Authorization header using the Bearer scheme.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}